Fix negative transform (#1383)
## Description

This PR fixes lightbox example(there was a problem with `runOnJS`). 

It also introduces a fix to a bug which occurred on Android(may also appear on ios but such situation hasn't been encountered). The problem was when animation transform received negative numbers, the regex didn't work properly(putting `-` sign as a prefix instead of treating it as a part of the number).

In the meantime I found a bug in animations on the web which will be fixed in this PR as well.