fix scheduler (#1473)
## Description

Currently, we use scheduler's instance variable `module` in order to recognize if scheduling still makes sense ().  However, if the scheduler has been cleaned we are actually accessing the field of the freed object. In order to get rid of the problem, I've changed the code so we capture module variable instead of the pointer to scheduler instance.

fixes: https://github.com/software-mansion/react-native-reanimated/issues/1459

## Changes

Capture a local copy of weak module ref. (REAIOSScheduler.mm)

## Test code and steps to reproduce

Rapidly change sth in worklets. (hard to reproduce)

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [x] Added TS types tests
- [x] Added unit / integration tests
- [x] Updated documentation
- [x] Ensured that CI passes