Don't run observer callbacks during sync uimanager batch update (#1501)
## Description

This PR fixes the way we perform sync updates on iOS by avoiding UIManager observers from firing. This in turn fixes the problem of ui manager commands that may be executed out of order as reported in #1498. The reason behind that issue was that due to the fact we'd trigger ui manager observers, those may enqueue some commands in the ui-manager block and we'd run them along the sync ui updates that we were expecting to get after layout phase. In the sync layout call we only expect the ui commands to come from the layout calculations and not from other sources (e.g. native animated or reanimated which uses ui manager observers to enqueue additional commands).

## Changes

1) This PR wraps uimanager batchDidComplete call with additional code that temporarily removes registered ui observers and only puts a single observer instead that is then used to capture mounting block.
2) Some additional cleanup in NodesManager on iOS to extract code that deals with RCTUIManager implementation details to a category on RCTUIManager.
3) As a part of this we also make an update to Android where we trigger semaphore release earlier when sync layout is not possible to match the implementation on iOS.
4) Opt-out of synchronous updates when no reanimated 2 code is being used by adding a flag `trySynchronousBatch` which is only set for reanimated 2 code paths. Then use this flag in the update procedure and skip semaphor wait on iOS and Android when it isn't set

### Before

Under certain condition (difficult to reproduce on a small app) reanimated 1 commands would sometimes run out of order (e.g. first "connect node" and only then "create node")

### After

We no longer allow for reanimated or native animated modules to enqueue commands that are goign to be executed in the sync layout phase.

## Test code and steps to reproduce

Apparently, we don't have a good repro steps for this, will test this internally on an app where we could reproduce the issue, and also reach out for help with testing to people who reported this in #1318