Fix scroll animation in Example (#1607)
## Description

The list in scroll Example app is inertness, and when list is moved and someone taps on screen and doesn't move by finger it calls only onStart() event from GestureHandler, and this sets the actual position but the list is still moving. If in the next step someone moves their finger then the list jumps to this position and starts animation of scroll from this position like in issue:

Fixes https://github.com/software-mansion/react-native-reanimated/issues/1597

## Changes
I just set the actual position of the animation to stop inertness of list.