Add previous result to useAnimatedReaction (#1653)
## Description

In some cases, we don't want to run the `reaction` callback from `useAnimatedReaction` if the value received from the `preparing` callback doesn't change. This equality could be checked always and calling the `reaction` callback would be then conditional. However, it's better to leave the decision to the user as there are as well cases when we still want to run the `reaction` even if its argument doesn't change.

To solve the problem, keeping backward compatibility, the value of the previous call of the `preparing` callback is now being passed to the `reaction`.

Besides the cases that were mentioned here, it lets users perform more custom operations.

## Code example

```Js
const x = useSharedValue(0);
const flag = useSharedValue(false);
useAnimatedReaction(
  () => {
    return x.value > 125 && flag.value;
  },
  (data, previous) => {
    if (data !== previous) {
		// only if it changes, perform some actions
    }
  }
);
```

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [x] Added TS types tests
- [x] Added unit / integration tests
- [x] Updated documentation
- [x] Ensured that CI passes