Shared value rerender behavior (#1660)
## Description

This is yet one more change to the logic of the `useSharedValue` hook and it comes as a modification of #1042. We decided to bring the rerender behavior of the `useSharedValue` closer to React's `useState` - every time a rerender happens(a state is changed), a shared value is not being rebuilt and it keeps the same value.
 
Fixes #1655 


#### Note

With that change, when having something like this `useSharedValue(x)` and changing the code that the hook receives a different value `useSharedValue(y)` it's not going to work with a fast refresh. This shared value will not have an updated value. Still, we decided to implement it that way as it's more important how it acts in the release mode than in the debug mode.

## Example code

```
import { useSharedValue } from 'react-native-reanimated';
import { View, Button } from 'react-native';
import React, { useState } from 'react';

export default function AnimatedStyleUpdateExample(props) {
  const [state, setState] = useState(10);
  const [stateCopy, setStateCopy] = useState(state);
  const sv = useSharedValue(state);

  setImmediate(() => {
    console.log(
      'render',
      state,
      sv.value,
      stateCopy
    );
  })

  return (
    <View>
      <Button
        title="set state"
        onPress={() => {
          setState(state + 1);
        }}
      />
    </View>
  );
}
```

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [x] Added TS types tests
- [x] Added unit / integration tests
- [x] Updated documentation
- [x] Ensured that CI passes