Fix ios crash (#2147)
fixes: https://github.com/software-mansion/react-native-reanimated/issues/2092
As you can see in the stack trace from the issue static object are deallocated before StoreUser destructor is called. 
It seems like the problem only occurs when a user closes an app. However, it's still a serious problem as crashing apps are treated differently on iOS. In order to get rid of the problem, I store old static members as one object which is kept as shared_ptr instance field.