add REAInitializer files (#2223)
If somebody has to turn off autoinstallation, they have to copy-paste our implementation of jsExecutorFactoryForBridge and update that code whenever we change something in the library. To make it a bit more flexible I want to add another layer of abstraction so that our code is hidden behind the REAJSIExecutorRuntimeInstaller function.