Fixed removing native views without ViewManager (#2486)
## Description

Some native Views don't have their own ViewManager. We had a check for this but this check was wrong. We did:
```java
ViewManager vm = nativeViewHierarchyManager.resolveViewManager(tag);
if (vm != null) {
    // code ...
}
```
But if it is unable to find view manager for tag `resolver` throws an exception instead of return null. I replace this check with `try {} catch() {}`

Fixes #2457

## Checklist

- [x] Ensured that CI passes