Add support for animating transform matrices (#2511)
## Description

This change adds support for animation transform matrices. Before, it was only possible to animate attributes that are either numbers or ones representing colors (e.g. `rgba(r,g,b,a)` or `#eae`). In react-native it is possible to not only provide transform as a list of operations like `translate` or `scale` but also by sending the full affine transform matrix (4x4 matrix represented as an array of 16 elements). This change makes it possible for reanimated to recognize such matrix and animate its transforms accordingly.

## Changes

 - added a way to recognize values of type array in utils.ts
 - override animation.onStart and animation.onFrame for the array type
 - copy animation to each individual element of an array and perform these separately (this isn't likely the most optimal approach as we are spanning 16 animation instances for a single matrix but this is a relatively rare usecase and this solution should be ok for the time being)
 - add new test app that uses inline transforms and also matrix transform to showcase the changes

## Test code and steps to reproduce

See newly added example `Example/test/TestApp.js` that demonstrate that changes made to matrix transform can be animated.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Ensured that CI passes