Remove _bridge_renimated, fix #2555 (#2565)
## Description

Fixes #2555. 

## Changes

Perhaps I'm missing something, but it seems to me that `_bridge_renimated` is a vestigial organ that no longer serves any practical purpose. Its use in REAEventDispatcher is unnecessary, since we already assign RCTEventDispatcher.bridge with the same value, and its only other use in createReanimatedModule is a single call where we can pass bridge as an argument.

Since bridge is now passed as an explicit dependency, a race condition is no longer possible.

I've realized that I can reproduce #2555 quite easily by repeatedly ⌘Ring in the simulator in quick succession. I've easily managed to get to the point where REAEventDispatcher has a null `_bridge_reanimated`, but `[self bridge]` is correct, confirming that the fix works.

## Screenshots / GIFs

n/a

## Test code and steps to reproduce

Reload the app in simulator repeatedly and see if gestures break as per #2555.

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [ ] Ensured that CI passes