Remove require on web (#3271)
## Description

`require('react-native/Libraries/Renderer/shims/ReactFabric').findHostInstance_DEPRECATED;` is causing a bunch of errors on web. I believe that webpack resolves require statically, so the surrounding `try .. catch` doesn't help.

## Changes

- adds a separate `fabricUtils.web.ts` file that doesn't try to load `findHostInstance_DEPRECATED` function, since Fabric is not supported on web at the moment

## Test code and steps to reproduce

I tested it on `reanimated-2-playground` with updated dependencies.
Running it on web with 3.0.0-rc.0 ends with a lot of errors. After this change the errors should be gone.

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [ ] Ensured that CI passes